Add files via upload
Imagine having to manage a small shop and having to work shifts 
every week. you have 4 employees N, V, G, D. 
The store has a 12-hour shift from 8:00 am and 8:00 pm over 7 days. 

N and V have a contract of 24 hours per week, while G,D of 40.
N,V can work 4 hours a day for 6 days 
or 5 hours a day for 4 days plus a day of 4.

G,D can work 6 hours a day for 4 days and 8 hours a day 
with 1 hour off for the remaining 2.

All 4 cannot work more than 8 consecutive days and each 
of them must have one free every 3 Sundays worked.

N is a working student so he prefers to work in the evening and more on weekends

V prefers to work in the morning to spend the afternoon with his children

D is a Napoli fan and therefore wants to see all his games 
(Otherwise he watches them while he works)

We want to create hours that comply with the laws imposed by the CCNL of Commerce 
but that are the best possible for employees